Adress reviewer comments for minimal_subscriber
