+++ rclcpp/topics/minimal_subscriber/CMakeLists.txt
@@ -32,14 +32,14 @@ ament_target_dependencies(subscriber_member_function_with_unique_network_flow_en
 add_executable(subscriber_not_composable not_composable.cpp)
 ament_target_dependencies(subscriber_not_composable rclcpp std_msgs)
 
-add_executable(subscriber_wait_set wait_set.cpp)
-ament_target_dependencies(subscriber_wait_set rclcpp std_msgs)
+add_executable(wait_set_subscriber wait_set_subscriber.cpp)
+ament_target_dependencies(wait_set_subscriber rclcpp std_msgs)
 
-add_executable(subscriber_static_wait_set static_wait_set.cpp)
-ament_target_dependencies(subscriber_static_wait_set rclcpp std_msgs)
+add_executable(static_wait_set_subscriber static_wait_set_subscriber.cpp)
+ament_target_dependencies(static_wait_set_subscriber rclcpp std_msgs)
 
-add_executable(subscriber_wait_set_time_triggered wait_set_time_triggered.cpp)
-ament_target_dependencies(subscriber_wait_set_time_triggered rclcpp std_msgs)
+add_executable(time_triggered_wait_set_subscriber time_triggered_wait_set_subscriber.cpp)
+ament_target_dependencies(time_triggered_wait_set_subscriber rclcpp std_msgs)
 
 install(TARGETS
   subscriber_lambda
@@ -48,9 +48,9 @@ install(TARGETS
   subscriber_member_function_with_type_adapter
   subscriber_member_function_with_unique_network_flow_endpoints
   subscriber_not_composable
-  subscriber_wait_set
-  subscriber_static_wait_set
-  subscriber_wait_set_time_triggered
+  wait_set_subscriber
+  static_wait_set_subscriber
+  time_triggered_wait_set_subscriber
   DESTINATION lib/${PROJECT_NAME})
 
 if(BUILD_TESTING)
